Rename desktop-notifications -> activity-notifications
