+++ shell/imports/server/activity-notifications.js
